requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.67
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
neo4j==5.28.1
python-dotenv==1.1.1
